+++ js/gen/app.post.js
@@ -6,10 +6,14 @@ app.post = {
 		// if retry_count is not defined......which will be the case for all callers of app.plugins.fetch
 		if (typeof retry_count == 'undefined') retry_count = 5;
 		var _data	= JSON.parse(JSON.stringify(data));
+
 		var url		= data._url || app.config.get_sisbot_url();
 		var timeout = 30000;
 
-		if (data.endpoint)		url		+= data.endpoint;
+		if (data.endpoint)		{
+			if (url.substr(-1) == '/' && data.endpoint.substr(0,1) == '/') url.replace(/\/$/, '');
+			url		+= data.endpoint;
+		}
 		if (data._timeout)		timeout = data._timeout;
 
 		var type = data._type || 'POST';
@@ -18,11 +22,11 @@ app.post = {
 		delete data._timeout;
 		delete data.endpoint;
 
+		console.log("Endpoint: "+data._url+" :: "+app.config.get_sisbot_url()+" = "+url+", Retry count " + retry_count);
+
 		var req_data = {
 			data	: JSON.stringify(data)
 		};
-
-		console.log("Retry count", retry_count);
 
 		if (app.current_user()) req_data.user = app.current_user().get('data');